chore(*) Cleaned up unused imports and other fixes

+++ src/processor/src/number.rs
@@ -109,7 +109,7 @@ impl Data {
             Self::Byte(value) => value as u64,
             Self::Word(value) => value as u64,
             Self::Dual(value) => value as u64,
-            Self::Quad(value) => value as u64
+            Self::Quad(value) => value
         }
     }
